Removed @Cached annotation and the complete caching feature
Caching added an unnecessary overhead of complexity to the framework which in the end didn't help as much as expected.
The number of milliseconds saved by the feature were minimal and in some cases the caching itself became an overhead.
Always resolving the WebElement anew is the way to go.